internal(vx-shape/SplitLinePath): remove dir structure, fix types, fewer multiplications, improve path measurement node attributes
